kafka: fix bugprone narrowing conversion
model::packed_record_batch_header_size is less than 60 bytes
so casting it to signed type won't risk any overflow.

+++ src/v/kafka/protocol/batch_reader.cc
@@ -35,7 +35,8 @@ public:
 
     int32_t size_bytes() const { return _size_bytes; }
     int32_t record_bytes() const {
-        return _size_bytes - model::packed_record_batch_header_size;
+        return _size_bytes
+               - static_cast<int32_t>(model::packed_record_batch_header_size);
     }
     model::offset last_offset() const {
         return _base_offset + model::offset{_last_offset_delta};
